+++ projects/website/src/app/components/layouts/dashboard/dashboard.html
@@ -1,7 +1,8 @@
 <header>
   <app-navbar></app-navbar>
 </header>
-
-<router-outlet />
+<section class="container">
+  <router-outlet />
+</section>
 
 <footer>Pied de page</footer>
