+++ projects/website/src/app/components/layouts/dashboard/dashboard.scss
@@ -1,0 +1,4 @@
+.container {
+  display: flex;
+  width: 100%;
+}
