+++ projects/website/src/app/components/shared/card/card.html
@@ -1,0 +1,3 @@
+<div class="card">
+  <ng-content></ng-content>
+</div>
