+++ projects/website/src/app/components/shared/card/card.scss
@@ -1,0 +1,6 @@
+.card {
+  display: flex;
+  padding: 20px;
+  background-color: rgb(255, 255, 255);
+  width: 60%;
+}
