+++ projects/website/src/app/components/shared/card/card.ts
@@ -1,0 +1,11 @@
+import { Component, input } from '@angular/core';
+
+@Component({
+  selector: 'app-card',
+  standalone: true,
+  templateUrl: './card.html',
+  styleUrl: './card.scss',
+})
+export class Card {
+  title = input<string>('');
+}
