+++ projects/website/src/app/components/shared/exam-setup/exam-setup.scss
@@ -1,0 +1,3 @@
+:host {
+  display: block;
+}
