+++ projects/website/src/app/components/shared/form/form.scss
@@ -1,0 +1,3 @@
+:host {
+  display: block;
+}
