--- projects/website/src/app/components/shared/input/input.html
@@ -1,3 +1,0 @@
-<h1 class="text-p-large"><ng-content></ng-content></h1>
-
-<input type="text" />
