--- projects/website/src/app/components/shared/input/input.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-text-p-large',
-  imports: [],
-  templateUrl: './input.html',
-  styleUrl: './input.scss',
-})
-export class Input {}
