+++ projects/website/src/app/components/shared/label/label.scss
@@ -1,7 +1,3 @@
-:host {
-  display: block;
-}
-
 .label {
   padding: 5px;
   &__text {
