+++ projects/website/src/app/components/shared/questions/questions.scss
@@ -1,7 +1,3 @@
-:host {
-  display: block;
-}
-
 .question {
   margin: 20px;
